Refactor IBC message parsing
