+++ cosmwasm/enclaves/shared/contract-engine/src/contract_operations.rs
@@ -18,7 +18,7 @@ use crate::contract_validation::verify_block_info;
 use crate::contract_validation::{ReplyParams, ValidatedMessage};
 use crate::external::results::{HandleSuccess, InitSuccess, QuerySuccess};
 use crate::message::{is_ibc_msg, parse_message};
-use crate::reply_message::ParsedMessage;
+use crate::types::ParsedMessage;
 
 use crate::random::update_msg_counter;
