+++ cosmwasm/enclaves/shared/contract-engine/src/lib.rs
@@ -12,13 +12,16 @@ mod contract_validation;
 mod cosmwasm_config;
 mod db;
 mod errors;
+mod execute_message;
 pub mod external;
 mod gas;
+mod ibc_message;
 mod io;
 mod message;
+mod message_utils;
 mod query_chain;
 mod random;
-pub mod reply_message;
+mod reply_message;
 pub(crate) mod types;
 #[cfg(feature = "wasmi-engine")]
 mod wasm;
